+++ src/core/GetLiveChatConfigOptionalParams.ts
@@ -1,0 +1,6 @@
+interface GetLiveChatConfigOptionalParams {
+    sendCacheHeaders?: boolean;
+    useRuntimeCache?: boolean;
+}
+
+export default GetLiveChatConfigOptionalParams;