+++ src/core/InitializeOptionalParams.ts
@@ -1,0 +1,7 @@
+import GetLiveChatConfigOptionalParams from "./GetLiveChatConfigOptionalParams"
+
+interface InitializeOptionalParams {
+    getLiveChatConfigOptionalParams?: GetLiveChatConfigOptionalParams;
+}
+
+export default InitializeOptionalParams